Refactor: move submodules to `src/` module
